final commit

+++ app/Http/Controllers/AdminProductBrandController.php
@@ -2,7 +2,9 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Product;
 use App\Models\ProductBrand;
+use App\Models\ProductCat;
 use Illuminate\Http\Request;
 
 class AdminProductBrandController extends Controller
@@ -10,19 +12,14 @@ class AdminProductBrandController extends Controller
     //
     function list()
     {
-        $product_brands = ProductBrand::paginate(5);
-        return view('admin.product.brand_list', compact('product_brands'));
-    }
-
-    function add()
-    {
-        return view('admin.product.brand_list');
+        $product_brands = ProductBrand::orderBy('created_at', 'desc')->paginate(5);
+        return view('admin.brands.list', compact('product_brands'));
     }
 
     function create(Request $request)
     {
         $request->validate([
-            'name' => 'required|string|max:100|unique:product_brands',
+            'name' => 'required|string|max:100|min:2|unique:product_brands',
         ], [
             'required' => ":attribute không được để trống",
             'min' => ":attribute phải có độ dài ít nhất :min kí tự",
@@ -42,30 +39,37 @@ class AdminProductBrandController extends Controller
 
     function delete($id)
     {
+        $products = Product::where('brand_id', '=', $id)->get();
+        
+        if ($products->count() > 0) {
+            return redirect('admin/product_brand/list')->with('status-danger', 'Không thể xóa thương hiệu này!');
+        }
+        
         ProductBrand::find($id)->delete();
+
         return redirect('admin/product_brand/list')->with('status', 'Xóa thương hiệu sản phẩm thành công!');
     }
 
     function update($id)
     {
         $product_brand = ProductBrand::find($id);
-        return view('admin.product.update_brand', compact('product_brand'));
+        return view('admin.brands.update', compact('product_brand'));
     }
 
     function store(Request $request, $id)
     {
         $request->validate([
-            'name' => 'required|string|max:100',
+            'name' => 'required|string|max:100|min:2',
             'status' => 'required',
         ], [
             'required' => ":attribute không được để trống",
             'min' => ":attribute phải có độ dài ít nhất :min kí tự",
             'max' => ":attribute phải có độ dài tối đa :max kí tự",
-
         ], [
             'name' => "Tên danh mục",
             'status' => "Trường trạng thái"
         ]);
+
 
         ProductBrand::where('id', $id)->update([
             'name' => $request->get('name'),
